+++ mockredis/redis.py
@@ -9,7 +9,7 @@ import string
 import sys
 
 from mockredis.lock import MockRedisLock
-from mockredis.exceptions import RedisError
+from mockredis.exceptions import RedisError, ResponseError
 from mockredis.pipeline import MockRedisPipeline
 from mockredis.script import Script
 from mockredis.sortedset import SortedSet
@@ -39,9 +39,9 @@ class MockRedis(object):
         self.strict = strict
         # The 'Redis' store
         self.redis = defaultdict(dict)
+        self.timeouts = defaultdict(dict)
         # The 'PubSub' store
         self.pubsub = defaultdict(list)
-        self.timeouts = defaultdict(dict)
         # Dictionary from script to sha ''Script''
         self.shas = dict()
 
@@ -165,12 +165,16 @@ class MockRedis(object):
         """
         Returns time to live in milliseconds if output_ms is True, else returns seconds.
         """
+        if key not in self.redis:
+            # as of redis 2.8, -2 returned if key does not exist
+            return long(-2)
         if key not in self.timeouts:
+            # redis-py returns None; command docs say -1
             return None
 
-        get_result = self._get_total_milliseconds if output_ms else self._get_total_seconds
+        get_result = get_total_milliseconds if output_ms else get_total_seconds
         time_to_live = get_result(self.timeouts[key] - currenttime)
-        return max(-1, time_to_live)
+        return long(max(-1, time_to_live))
 
     def ttl(self, key, currenttime=datetime.now()):
         """
@@ -255,7 +259,7 @@ class MockRedis(object):
                 expire = px if isinstance(px, timedelta) else timedelta(milliseconds=px)
 
             if expire is not None and expire.total_seconds() <= 0:
-                raise ValueError("invalid expire time in SETEX")
+                raise ResponseError("invalid expire time in SETEX")
 
             result = self._set(key, value)
             if expire:
@@ -593,11 +597,11 @@ class MockRedis(object):
         """Emulate lset."""
         redis_list = self._get_list(key, 'LSET')
         if redis_list is None:
-            raise ValueError("no such key")
+            raise ResponseError("no such key")
         try:
             redis_list[index] = value
         except IndexError:
-            raise ValueError("index out of range")
+            raise ResponseError("index out of range")
 
     #### SET COMMANDS ####
 
@@ -717,7 +721,7 @@ class MockRedis(object):
 
         # args
         if len(args) % 2 != 0:
-            raise ValueError("ZADD requires an equal number of "
+            raise RedisError("ZADD requires an equal number of "
                              "values and scores")
         for i in xrange(len(args) // 2):
             # interpretation of args order depends on whether Redis
@@ -792,7 +796,7 @@ class MockRedis(object):
     def zrangebyscore(self, name, min_, max_, start=None, num=None,
                       withscores=False, score_cast_func=float):
         if (start is None) ^ (num is None):
-            raise TypeError('`start` and `num` must both be specified')
+            raise RedisError('`start` and `num` must both be specified')
 
         zset = self._get_zset(name, "ZRANGEBYSCORE")
 
@@ -858,7 +862,7 @@ class MockRedis(object):
     def zrevrangebyscore(self, name, max_, min_, start=None, num=None,
                          withscores=False, score_cast_func=float):
         if (start is None) ^ (num is None):
-            raise TypeError('`start` and `num` must both be specified')
+            raise RedisError('`start` and `num` must both be specified')
 
         zset = self._get_zset(name, "ZREVRANGEBYSCORE")
         if not zset:
@@ -989,15 +993,6 @@ class MockRedis(object):
 
     #### Internal ####
 
-    def _get_total_seconds(self, td):
-        """
-        For python 2.6 support
-        """
-        return int((td.microseconds + (td.seconds + td.days * 24 * 3600) * 1e6) // 1e6)
-
-    def _get_total_milliseconds(self, td):
-        return int((td.days * 24 * 60 * 60 + td.seconds) * 1000 + td.microseconds / 1000.0)
-
     def _get_list(self, key, operation, create=False):
         """
         Get (and maybe create) a list by name.
@@ -1059,7 +1054,7 @@ class MockRedis(object):
         Return a suitable function from (score, member)
         """
         if withscores:
-            return lambda score_member: (score_member[1], score_cast_func(score_member[0]))
+            return lambda score_member: (score_member[1], score_cast_func(str(score_member[0])))
         else:
             return lambda score_member: score_member[1]
 
@@ -1078,7 +1073,7 @@ class MockRedis(object):
         """Helper function for sdiff, sinter, and sunion"""
         keys = self._list_or_args(keys, args)
         if not keys:
-            raise ValueError("wrong number of arguments for '{}' command".format(operation.lower()))
+            raise TypeError("{} takes at least two arguments".format(operation.lower()))
         left = self._get_set(keys[0], operation) or set()
         for key in keys[1:]:
             right = self._get_set(key, operation) or set()
@@ -1101,6 +1096,17 @@ class MockRedis(object):
         if args:
             keys.extend(args)
         return keys
+
+
+def get_total_seconds(td):
+    """
+    For python 2.6 support
+    """
+    return int((td.microseconds + (td.seconds + td.days * 24 * 3600) * 1e6) // 1e6)
+
+
+def get_total_milliseconds(td):
+    return int((td.days * 24 * 60 * 60 + td.seconds) * 1000 + td.microseconds / 1000.0)
 
 
 def mock_redis_client(**kwargs):